requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
Django==5.2.5
django-allauth==65.11.0
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.10
pycparser==2.22
PyJWT==2.10.1
python-decouple==3.8
psycopg2-binary==2.9.9
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
pytest==8.3.4
pytest-django==4.9.0
factory-boy==3.3.1
pytest-cov==6.0.0
coverage==7.6.7
reportlab==4.2.5
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.1.1
pytest-html==4.1.1
